requirements.txt
flask
quart
requests
openai
anthropic
google-generativeai
sqlalchemy
datasets
python-dotenv
pytest
matplotlib
plotly
pandas
seaborn
selenium
beautifulsoup4
transformers
torch
numpy
scikit-learn
aiohttp